+++ src/connection.ts
@@ -784,6 +784,57 @@ export class Connection {
   /**
    * @internal
    *
+   * Waits for propagation.
+   *
+   * See {@link Database.waitForPropagation}.
+   *
+   * @param request - Request to perform against each coordinator.
+   * @param timeout - Maximum number of milliseconds to wait for propagation.
+   */
+  async waitForPropagation(request: RequestOptions, timeout = Infinity) {
+    const numHosts = this._hosts.length;
+    const propagated = [] as number[];
+    const started = Date.now();
+    let host = 0;
+    while (true) {
+      if (propagated.length === numHosts) {
+        return;
+      }
+      while (propagated.includes(host)) {
+        host = (host + 1) % numHosts;
+      }
+      try {
+        await this.request({ ...request, host });
+      } catch (e) {
+        if (started + timeout < Date.now()) {
+          console.error(
+            request.path,
+            `Timeout after ${((Date.now() - started) / 1000).toFixed(
+              2
+            )} secs with ${propagated.length} of ${numHosts} propagated!`,
+            String(e)
+          );
+          throw e;
+        }
+        await new Promise((resolve) => setTimeout(resolve, 1000));
+        continue;
+      }
+      if (!propagated.includes(host)) {
+        console.error(
+          request.path,
+          `Host ${host} propagated, ${
+            numHosts - (propagated.length + 1)
+          } remaining.`,
+          `${((Date.now() - started) / 1000).toFixed(2)} secs elapsed...`
+        );
+        propagated.push(host);
+      }
+    }
+  }
+
+  /**
+   * @internal
+   *
    * Performs a request using the arangojs connection pool.
    */
   request<T = ArangojsResponse>(