+++ src/test/09-collection-metadata.ts
@@ -15,6 +15,10 @@ describe("Collection metadata", function () {
     await db.createDatabase(dbName);
     db.useDatabase(dbName);
     collection = await db.createCollection(collectionName);
+    await db.waitForPropagation(
+      { path: `/_api/collection/${collection.name}` },
+      30000
+    );
   });
   after(async () => {
     db.useDatabase("_system");