+++ src/test/13-bulk-imports.ts
@@ -14,6 +14,10 @@ describe("Bulk imports", function () {
     await db.createDatabase(dbName);
     db.useDatabase(dbName);
     collection = await db.createCollection(collectionName);
+    await db.waitForPropagation(
+      { path: `/_api/collection/${collection.name}` },
+      30000
+    );
   });
   after(async () => {
     try {