+++ src/test/14-document-collections.ts
@@ -24,6 +24,10 @@ describe("DocumentCollection API", function () {
   });
   beforeEach(async () => {
     collection = await db.createCollection(`c_${Date.now()}`);
+    await db.waitForPropagation(
+      { path: `/_api/collection/${collection.name}` },
+      30000
+    );
   });
   afterEach(async () => {
     await collection.drop();