+++ src/test/19-graph-vertex-collections.ts
@@ -21,6 +21,7 @@ describe("GraphVertexCollection API", function () {
         to: ["person"],
       },
     ]);
+    await db.waitForPropagation({ path: `/_api/gharial/${graph.name}` }, 30000);
     collection = graph.vertexCollection("person");
   });
   after(async () => {