+++ src/test/29-manipulating-analyzers.ts
@@ -3,12 +3,6 @@ import { Analyzer } from "../analyzer";
 import { Database } from "../database";
 import { config } from "./_config";
 
-async function sleep(ms: number) {
-  return new Promise<void>((resolve) => {
-    setTimeout(() => resolve(), ms);
-  });
-}
-
 const describe35 = config.arangoVersion! >= 30500 ? describe : describe.skip;
 
 describe35("Manipulating analyzers", function () {
@@ -33,6 +27,10 @@ describe35("Manipulating analyzers", function () {
       const analyzer = db.analyzer(`a_${Date.now()}`);
       expect(await analyzer.exists()).to.equal(false);
       await analyzer.create({ type: "identity" });
+      await db.waitForPropagation(
+        { path: `/_api/analyzer/${analyzer.name}` },
+        120000
+      );
       expect(await analyzer.exists()).to.equal(true);
     });
   });
@@ -41,6 +39,10 @@ describe35("Manipulating analyzers", function () {
     before(async () => {
       analyzer = db.analyzer(`a_${Date.now()}`);
       await analyzer.create({ type: "identity" });
+      await db.waitForPropagation(
+        { path: `/_api/analyzer/${analyzer.name}` },
+        120000
+      );
     });
     after(async () => {
       try {
@@ -56,6 +58,10 @@ describe35("Manipulating analyzers", function () {
     it("creates the analyzer", async () => {
       const analyzer = db.analyzer(`a_${Date.now()}`);
       await analyzer.create({ type: "identity" });
+      await db.waitForPropagation(
+        { path: `/_api/analyzer/${analyzer.name}` },
+        120000
+      );
       const data = await analyzer.get();
       expect(data).to.have.property("name", `${name}::${analyzer.name}`);
       expect(data).to.have.property("type", "identity");
@@ -65,7 +71,10 @@ describe35("Manipulating analyzers", function () {
     it("destroys the analyzer", async () => {
       const analyzer = db.analyzer(`a_${Date.now()}`);
       await analyzer.create({ type: "identity" });
-      await sleep(3000);
+      await db.waitForPropagation(
+        { path: `/_api/analyzer/${analyzer.name}` },
+        120000
+      );
       await analyzer.drop();
       expect(await analyzer.exists()).to.equal(false);
     });